requirements.txt
python-telegram-bot[job-queue]==21.6
openai==1.51.0
gspread==6.1.2
google-auth==2.35.0
httpx==0.27.2
